fix: 로그인 전역값을 활용한 silentRefresh 이후 api 실행
