+++ src/pages/common/paymentStep/Step.tsx
@@ -5,11 +5,6 @@ import { useSetAtom } from 'jotai';
 import { navigateNextStepAtom } from '~/models/funnel';
 import { useStepToGoBack } from '~/hooks/useStepToGoBack';
 import useTypeSafeNavigate from '~/hooks/useTypeSafeNavigate';
-import { PaymentAPI } from '~/api';
-import toast from 'react-hot-toast';
-import { useEffect, useState } from 'react';
-import { PaymentResponse } from '~/api/types/payment.type';
-import axios from 'axios';
 
 const PAGE_NUMBER = [1];
 
@@ -25,27 +20,6 @@ const CommonPaymentStep = () => {
   const stepToGoBackGroup = useStepToGoBack('groupLeaderPledgeStep');
   const stepToGoBack = stepToGoBackPersonal && stepToGoBackGroup;
   const navigate = useTypeSafeNavigate();
-  const [userPaymentInfo, setUserPaymentInfo] = useState<PaymentResponse>();
-  const handlePaymentRequest = async () => {
-    try {
-      const res = await PaymentAPI.requestPayment({
-        pg: 'WELCOME_PAYMENTS',
-        payMethod: 'card',
-      });
-      setUserPaymentInfo(res.data);
-    } catch (error) {
-      console.log(error);
-      if (axios.isAxiosError(error) && error.response?.data.code === 'P04') {
-        toast.success('이미 신청하셨습니다!', {
-          duration: 1800,
-        });
-      }
-    }
-  };
-
-  useEffect(() => {
-    setTimeout(handlePaymentRequest, 500);
-  }, []);
 
   if (stepToGoBack) {
     setNavigateNextStep(true);
@@ -58,7 +32,7 @@ const CommonPaymentStep = () => {
       <PageLayout.SingleCardBody cardPadding={'0'}>
         <Funnel>
           <Funnel.Page pageNumber={1}>
-            <PaymentPage userPaymentInfo={userPaymentInfo} />
+            <PaymentPage />
           </Funnel.Page>
         </Funnel>
       </PageLayout.SingleCardBody>
